livekit-voice-agent/pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents[silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation>=0.2",
    "python-dotenv>=1.2.1",
    "torch>=2.9.1",
    "transformers>=4.57.1",
    "tzdata>=2025.2",
]